+++ AlchemistCode/Cards/Common/Puncture.cs
@@ -13,7 +13,7 @@ public class Puncture : AlchemistCard
 {
     public Puncture() : base(1, CardType.Attack, CardRarity.Common, TargetType.AnyEnemy)
     {
-        WithDamage(6, 2);
+        WithDamage(7, 2);
         WithPower<VulnerablePower>(2, 1);
         WithTip(typeof(PoisonPower));
     }