+++ AlchemistCode/Cards/Rare/Croak.cs
@@ -14,7 +14,6 @@ public class Croak : AlchemistCard
     public Croak() : base(2, CardType.Attack, CardRarity.Rare, TargetType.AnyEnemy)
     {
         WithDamage(20, 6);
-        WithKeyword(CardKeyword.Exhaust);
         WithTip(StaticHoverTip.Fatal);
     }